requirements.txt
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
cycler==0.10.0
decorator==4.4.2
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram==0.12.0
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
tornado==6.0.4
urllib3==1.25.10
